1.修改完成菜单数据初始化参数功能，能够正常读取EEPROM数据，写入默认参数，

+++ demo/Project/Project_template/inc/flash.h
@@ -69,7 +69,7 @@ typedef struct _FACTORY_DATA_
   uint16_t  overmoment2;				//过力矩系数2
   uint16_t  in4ma;						//输入4ma校准
   uint16_t  in20ma;						//输入20ma校准
-  uint16_t  *number;					//产品编号地址
+  uint8_t  *number;						//产品编号地址
   uint16_t  factor;						//系数
   uint16_t  overdelay;					//过力矩延时
 }FactoryDataStr;
